+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Updated wsjcpp-core to v0.2.3 (Up C++ from 11 to 17)
 - Change single name registartion to multiple names
+- Added silent mode
+- Added class for simplify init `WsjcppEmployeesInit`
 
 ### Added